Finished basic map generation
Program now generates a map of specified tiles with the variables present in Game Manager.
Script needs to be modified a bit in the future to prevent tiles from overlapping, there is already something in place to help reduce overlapping, but it checks tiles sequentially, which means in moving to avoid overlapping a later tile, it can overlap an earlier one.